Cleanup for data warehouse commands.
Options are adjusted to be more standard. The file name for an import is
now optional, standard input is read when no file is supplied. The
import command now displays a summary of the imported data which
includes the number of events and the date ranges found in the export
file.

There is now no logging by default, a new "-E" option allows error
logging to be enabled.

It is now possible to pass options through to the JVM. This allows
customization of memory or other settings.