Commands updated for better handling of passwords and error output.
The database password is now passed from Python to Java in the
environment rather than on the command line.

There is no longer a default value for the database password. If a
password is not supplied (and a console is available) then the user is
prompted to enter the database password.

Logging and error output is now directed to standard error (was to
standard out)