Commands updated to support SSL/TLS database connections.
A new database option is added to allow the SHA-1 fingerprint of the
database servers SSL certificate. Either the full fingerprint can be
supplied or a prefix, the empty string will allow use of any server
certificate.

The defaults for the SSL/TLS protocol and enabled cipher suites are
reasonable. If desired these values can be overridden using a
configuration file or the (hidden) command line options.

If connections to the database fail due to an untrusted certificate then
the issuer/serial, subject and SHA-1 fingerprint of the server
certificate are displayed to allow verification and copying of the
fingerprint to the configuration file.